#3871 Extract 'Hosted mode' in separate extension
