+++ examples/electron/package.json
@@ -39,7 +39,8 @@
     "@theia/navigator": "^0.7.0",
     "@theia/outline-view": "^0.7.0",
     "@theia/output": "^0.7.0",
-    "@theia/plugin": "^0.7.0",
+    "@theia/plugin-dev": "^0.7.0",
+    "@theia/plugin-ext": "^0.7.0",
     "@theia/plugin-ext-vscode": "^0.7.0",
     "@theia/preferences": "^0.7.0",
     "@theia/preview": "^0.7.0",
