+++ packages/plugin-dev/README.md
@@ -1,6 +1,9 @@
-# Theia - Theia - Plugin API
+# Theia - Plugin Development
 
 See [here](https://www.theia-ide.org/doc/index.html) for a detailed documentation.
+
+## Contribution points:
+ - Hosted Instance uri post processor
 
 ## License
 - [Eclipse Public License 2.0](http://www.eclipse.org/legal/epl-2.0/)
