+++ packages/plugin-dev/src/browser/hosted-plugin-informer.ts
@@ -18,7 +18,7 @@ import { injectable, inject } from 'inversify';
 import { StatusBar } from '@theia/core/lib/browser/status-bar/status-bar';
 import { StatusBarAlignment, StatusBarEntry, FrontendApplicationContribution } from '@theia/core/lib/browser';
 import { WorkspaceService } from '@theia/workspace/lib/browser';
-import { HostedPluginServer } from '../../common/plugin-protocol';
+import { HostedPluginServer } from '../common/plugin-dev-protocol';
 import { ConnectionStatusService, ConnectionStatus } from '@theia/core/lib/browser/connection-status-service';
 import URI from '@theia/core/lib/common/uri';
 import { FileStat } from '@theia/filesystem/lib/common';
