+++ packages/plugin-dev/src/browser/hosted-plugin-preferences.ts
