+++ packages/plugin-dev/src/node/hosted-plugin-uri-postprocessor.ts
